builder/requirements.txt
PyPDF2
langchain
openai
faiss-cpu
tiktoken
python-dotenv
flask